Serde-encode info data instead of generating code.
This encodes the register tables parsed from the Mesa sources into a
serialized data structure, which is then embedded in the rodata of the
vsc7448-info crate and deserialized on-demand.

Currently I'm using postcard, the same codec Bryan uses in portions of
Humility, because it's faster and smaller than bincode. I experimented
with using human-readable formats like RON, but, they slow.

This reduces release compile time for the vsc7448-info crate on my Ryzen
7 laptop from 1m42s to ~5s. Runtime parsing of the serialized tables
takes about 5ms (in a release build, 46ms in a debug build).

+++ mesa-parse/src/symregs.rs
@@ -13,7 +13,7 @@ use vsc7448_types::Address;
 pub type TargetMap = BTreeMap<String, (Address, BTreeMap<String, Address>)>;
 
 // Represents the layout of Targets in memory.
-pub type TargetList = BTreeMap<String, (String, Vec<(Option<usize>, usize)>)>;
+pub type TargetList = BTreeMap<String, (String, Vec<(Option<u32>, u32)>)>;
 
 type MemoryMap = (BTreeMap<String, TargetMap>, TargetList);
 
@@ -112,16 +112,12 @@ pub fn parse_symregs(s: &str) -> MemoryMap {
         if let Some(caps) = target_re.captures(s) {
             let name = caps[1].trim_matches('\"').to_owned();
             let repl: i32 = caps[2].parse().unwrap();
-            let addr = parse_int::parse::<usize>(&caps[5]).unwrap();
+            let addr = parse_int::parse::<u32>(&caps[5]).unwrap();
             let entry = target_list
                 .entry(name)
                 .or_insert((caps[6].to_owned(), Vec::new()));
             assert!(entry.0 == caps[6]);
-            let repl = if repl == -1 {
-                None
-            } else {
-                Some(repl as usize)
-            };
+            let repl = if repl == -1 { None } else { Some(repl as u32) };
             entry.1.push((repl, addr + offsets.get(&caps[4]).unwrap()));
         }
     }
